+++ docs/release-notes.md
@@ -6,6 +6,7 @@
 
 ### For everyone:
 * Updated the Pintail library, improving framerate.
+* The artifact codex now has separate categories for ship-specific and event artifacts.
 
 ### For developers:
 * The `IModCharacters(.V2).LookupByUniqueName` and `IModShips.LookupByUniqueName` methods now return correct information for vanilla entries.